Proje Kodları/requirements.txt
google-generativeai==0.5.2
streamlit==1.34.0
sentence-transformers==2.2.2
huggingface-hub==0.14.1
faiss-cpu==1.7.4
pandas==2.2.2
torch==2.2.2
openpyxl==3.1.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
joblib==1.3.2
openai==0.28.1 